# profile_swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Member Profile System",
    "description": "Membership system for Fusion Products",
    "version": "1.0.0"
  },
  "host": "localhost",
  "schemes": [
    "https"
  ],
  "basePath": "/api",
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/member/{member_id}": {
      "get": {
        "summary": "Get a particular Member",
        "description": "The member endpoint returns a `Member` objects.\n",
        "parameters": [
          {
            "name": "key",
            "in": "header",
            "description": "The product key assigned to the product.",
            "required": true,
            "type": "string"
          },
          {
            "name": "secret",
            "in": "header",
            "description": "The secret assigned to the product to execute",
            "required": true,
            "type": "string"
          },
          {
            "name": "member_id",
            "in": "path",
            "description": "The member id",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Member"
        ],
        "responses": {
          "200": {
            "description": "A Member",
            "schema": {
              "$ref": "#/definitions/Member"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a member by product owner",
        "description": "Delete a member by product owner",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "description": "The member id",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Member"
        ],
        "responses": {
          "200": {
            "description": "Member deleted successfully."
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Member": {
      "type": "object",
      "properties": {
        "member_id": {
          "type": "string",
          "description": "Unique identifier representing a specific member for a given product."
        },
        "product_id": {
          "type": "string",
          "description": "Representing a product for which member is going to signup"
        },
        "username": {
          "type": "string",
          "description": "Member username."
        },
        "email": {
          "type": "string",
          "description": "Member email."
        },
        "password": {
          "type": "string",
          "description": "Member password."
        },
        "profile": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Profile_Property"
          },
          "description": "Customisable profile json."
        },
        "image": {
          "type": "string",
          "description": "Image URL representing the member."
        }
      }
    },
    "Profile_Property": {
      "type": "object",
      "properties": {
        "label": {
          "type": "string",
          "description": "The JSON label/key for the profile field"
        },
        "value": {
          "type": "string",
          "description": "The value of the JSON field"
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        }
      }
    },
...
  }
}